move TLDParser classes to their own file
